Move `PrimerUIManager`
